+++ extension/frontend/view/view.html
@@ -7,8 +7,8 @@
     <title>Test Recoil</title>
 </head>
 <body>
-   
-    <div id="total-recoil">Total RecoilJS</div>
+   <div id="root"></div>
+    <!-- <div id="total-recoil">Total RecoilJS</div>
     <div id="currentComponent" >No Atoms Hovered</div>
     <div id="main-container">
       <div id="tree"></div>
@@ -18,12 +18,12 @@
           <button>History</button>
         </div>
         <div class='information'></div>
-        <!-- <h1>Node Information</h1>
+         <h1>Node Information</h1>
         <h1 id="state-info"></h1> -->
-      </div>
+      <!-- </div>
     </div>
   
-  </div>
+  </div> -->
     <script src="bundle.js"></script>
 </body>
 </html>
