+++ package.json
@@ -12,6 +12,8 @@
   "license": "ISC",
   "devDependencies": {
     "@babel/preset-env": "^7.10.2",
+    "@babel/polyfill": "^7.10.4",
+    "@babel/preset-react": "^7.10.1",
     "babel-jest": "^26.0.1",
     "copy-webpack-plugin": "^6.0.1",
     "css-loader": "^3.5.3",
@@ -26,6 +28,11 @@
     "webpack-extension-reloader": "^1.1.4"
   },
   "dependencies": {
-    "lodash.throttle": "^4.1.1"
+    "babel-loader": "^8.1.0",
+    "d3": "^5.16.0",
+    "lodash.throttle": "^4.1.1",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-router-dom": "^5.2.0"
   }
 }
